Fix some classes organiation
